Implemented new approach for memleakfix.
Old code had a memory overflow bug and also new approach is cleaner.

Important: I left some code that needs to be cleaned up but shouldn't
interfere with new approach. I left this code for now as I was worried
that I may not remove it appropriately. More info in comments.